+++ lib/views/index.erb
@@ -12,12 +12,3 @@
     <input type='submit' value='Add peep'>
   </form>
 </div>
-
-<% if flash[:errors] && !flash[:errors].empty? %>
-  Sorry, there were the following problems with your registration:
-  <ul id='errors'>
-    <% flash[:errors].each do |error| %>
-      <li><%= error %></li>
-    <% end %>
-  </ul>
-<% end %>
